Animations
Added animations for the whale boss, barnacle, and puffer fish.  The
whale and puffer fish animations used a different model than the
existing prefabs, so they will need to have all of the scripts/effects
reapplied to them.  Also added 2 lines in LanderCollision.cs to
pause/play the landing animation.